Change #3.0
- Deleted “private” markers on variables, as they were already private
to begin with
- Replaced variables _sandPosX, _sandPosY, and _sandFallSpeed with
PVectors sandPosition & gravity as well as attempted to understand
vectors
- Removed the color changing from the Sand class, moving parts of it to
the main area (correct name?)
- Removed the MusicPlayer class entirely—better as just a regular
custom function rather than its own class, as it is very static and
only used once. Potential the same fate coming to the Square class some
time in the future
- Began the process of overhauling x & y coordinate locations.
Previously, these were similar integers, but starting to use terms
defined by width and height instead
- Attempted to fix collisions, failed yet again—“and I would have
gotten away with a good grade too, if it wasn’t for you pesky
collisions!”
- Switched color change to a simple “fill all” using a boolean. Will
change this to sand.add(new Sand) with a NEW COLOR variable sometime in
the future (need some help on this front)

Current Necessities:

- Get a better grasp on ArrayLists and Vectors/PVector, particularly in
terms of syntax
- Get collisions UP AND RUNNING FOR REAL
- Finish color changing
- Implement a settings menu (with options to adjust “wind
speed”—potential wind vector?—gravity, size, etc).
- Clean up code significantly
- Add comments
- Width/height-ify things to make it scale on all screens
- Implement a “back button” for the music player, as well as a way to
disable selection of the same song twice?
- Just do some work for crying out loud, sheesh

+++ Square.pde
@@ -1,6 +1,6 @@
 class Square
 {
-  private int _squarePosX, _squarePosY, _squareWidth, _squareHeight;
+  private float _squarePosX, _squarePosY, _squareWidth, _squareHeight;
   private color _col;
   Square(int squarePosX)
   {
